refactor: move utils into a separate package
