+++ packages/ouro-core/package.json
@@ -8,6 +8,7 @@
   "module": "dist/index.es.js",
   "dependencies": {
     "ouro-result": "^0.0.1",
-    "ouro-traits": "^0.0.1"
+    "ouro-traits": "^0.0.1",
+    "ouro-utils": "^0.0.1"
   }
 }
