--- packages/ouro-core/src/__tests__/utils.js
@@ -1,8 +1,0 @@
-// @flow
-
-import { identity } from '../utils'
-
-test('#identity()', () => {
-  const subj = {}
-  expect(identity(subj)).toBe(subj)
-})
