+++ packages/ouro-utils/src/__tests__/identity.js
@@ -1,0 +1,13 @@
+// @flow
+
+import identity from '../identity'
+
+let subj
+
+beforeEach(() => {
+  subj = {}
+})
+
+test('it returns the input argument', () => {
+  expect(identity(subj)).toBe(subj)
+})
