+++ packages/ouro-utils/src/identity.js
@@ -1,0 +1,5 @@
+// @flow
+
+export default function identity<T>(value: T): T {
+  return value
+}
