+++ packages/ouro-utils/src/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export { default as identity } from './identity'
+export { default as reduce } from './reduce'
